requirements.txt
requests>=2.25.1
pandas>=1.3.0
openpyxl>=3.0.7
matplotlib>=3.5.0
python-telegram-bot>=20.0
tqdm>=4.60.0
prettytable>=2.0.0
colorama>=0.4.0
python-dotenv>=0.19.0